requirements.txt
mcp==1.9.3
mcp[cli]==1.9.3
uv==0.7.11
huggingface_hub[mcp]>=0.32.0
gradio[mcp]==5.30.0
textblob==0.19.0
smolagents[mcp]==1.17.0
gradio_client==1.10.1
newsapi-python==0.2.7
fastmcp==2.3.0
yfinance==0.2.62
langchain-openai~=0.3.22
langgraph~=0.4.8
pandas~=2.3.0
pydantic~=2.11.5
pydantic-settings~=2.9.1